Generalizing curvature correction to all inst
